refactor: reorganize components and assets
